cleanup: move screens into pages and components in their subdirectories
