+++ src/components/home/widgets/SignInOutput.svelte
