+++ src/pages/Crash.svelte
@@ -1,5 +1,5 @@
 <script>
-  import TransitionWrapper2 from "./TransitionWrapper2.svelte";
+  import PopuotWindowTransitionWrapper from "./PopuotWindowTransitionWrapper.svelte";
   import VirtualList from "../components/utils/VirtualList.svelte";
   import LogMessage from "../components/log/LogMessage.svelte";
   import { onMount } from "svelte";
@@ -29,7 +29,7 @@
   }
 </script>
 
-<TransitionWrapper2>
+<PopuotWindowTransitionWrapper>
   <div class="logs-wrapper">
     <VirtualList items={$minecraftLogs} let:item {autoScroll}>
       <LogMessage text={item} />
@@ -41,7 +41,7 @@
       </h1>
     </div>
   </div>
-</TransitionWrapper2>
+</PopuotWindowTransitionWrapper>
 
 
 <style>
